Improved dependency sorting
Ordering is leveled now, opening up a simple
path for a naive paralellization of pre compiling.
Further improvements are needed to advance beyond a level before
all other modules on that level are done. For situations where
a higher level branch may reach a dependency before a slower,
but lower, level we need to bind nodes and prune.

To Be Continued...