--projects-file=, use a file instead of fetching
Normally fetching involves fetching the projects.json
file from that authority. This allows a user to do:
--projects-file=projects.json to instead from-json on
that file and use that instead.

For smoke testing we now filter on @ignore, then save
this filtered projects list to a file. For each 'install'
method called we can then pass in this generated file
(as each install is a new Proc)