Add :$pad arg; Improve :@alpha assignment
The padding character is changable via :$pad.

The encoding array can be be passed as an array of 2 or 64 values.
If it is only 2 values they get appended to the default 62 base64
standard characters. If an array of 64 *unique* values is passed
then the entire array will be used as the encoding chars.