Fixes for \r\n rakudo changes; removing http client
The recent \r\n changes in rakudo have left the http parsing
grammar unworking, and i'm unable to see how it can possibly
work with how \r\n is handled. In additiona, the grammar was
moved to Grammar::HTTP already. Since the http client is only
used for fetching a json file, and because I plan to change
this so you can just `git` the projects list instead, the http
client has started to be removed. It already resides at
Net::HTTP, so we can make it optional or something.