Better Proc fix for when we need to save them
In the previous fix we were only interested in the exit code.
Here however we want to keep failed processes around so we can
examine them later and report on them. If we just catch the
exception and assign the passed failed proc to a variable, it
rethrows the exception. To get around this we just shadow
the sink method to do nothing which gets us the old behavior
of Proc back