Redo win32 proc.lines fix to work with jvm as well
Proc.out contains "\n" for newlines even on windows. I expect
this is a bug and will be corrected, as its otherwise impossible
to .split("\n") as the \n will be turned into \r\n on windows.
This fix decodes ord '10' and uses that as a "\n" for split (which
prevents it getting turned into \r\n). Once the actual bug is
fixed I imagine this will break on windows until it is simply
removed.