Add Distribution::DependencySpecification
Zef::Distribution::DependencySpecification can be used to match
an identity string (as used in depends field of meta) to create
a matcher that can be used against a Distribution object. While
we assume the ContentStorage gives us correct results, this is
now used in the topological sort so we can sort an array of
Zef::Distribution objects using smart match instead of passing
the meta hash of all fetched dists and then trying to match the
various fields in multiple steps.