Murder the Build.pm hook shit fest
A giant fuck all to fix the sad state of build hooks because the lack of a documentated core interface
has left many writing code that only works for a single package manager when its not needed. Why
a package manager would force themselves as a dependency when it only wants to know if an interface
is fulfilled is beyond me so I have no qualms with modifying code to make it work proper.
What *SHOULD* happen is CompUnit::Repository::Installation and friends should define the damned
interfaces themselves so not only does it leave the ecosystem in a better state (less fragile dependency
chains) but reinstallation can also re-execute the hooks instead of (once again) demanding you have
a dependency on an external package manager. This is not to be construed as a jab against the author,
as Build.pm did what was needed *at the time*, but its 2016 and perl6 is released so lets do this shit right.