Handle fetching identities of dist or module
Now a dependency like `XXX:ver<1>:auth<xxx:fff>` gets treated
properly. Same with a dependency on a MODULE name in the same form
(not a dist name) such that requesting URI::Escape fetches from
dist URI. It does this by creating DependencySpecifications for
depends and provides of a Distribution object.