Fix 007 failure cases patrickz++
Fixes:
1. spec-matcher was backwards; now skips over a `return False`
when the auth is not part of the *requested* identity instead
of wanting to find a distro with a matching missing auth.

2. $dist.is-installed (which its current form is temporary anyway)
now checks if it can use $dist.name *or* $dist.provides[0].name.
The 007 dist has no module named '007' to EVAL('use 007'); So by
checking provides we get a better idea: EVAL('use _007::Runtime')

3. ::Test searches for test files recursively now by default, as
007 does not have any tests in the root of its /t directory

Thanks to patrickz in #perl6-toolchain for the bug report