Basic Messenger implementation
Give plugins a common target to send their stdout/stderr to that
we can control (i.e. we don't have to pass :quiet to every plugin,
just the phase itself which will pass in an appropriate tap/act).

Eventually it will pass a Messenger object (or hash?) to identity
the phase its from, the verbosity level, the file its processing,
and other such details that may be displayed differently depending
on parallelization, if its a CLI or GUI, etc.