Be more clear about needing --force
First, shell out to test if a module is installed so we can use
it on ourselves (the old EVAL('use Zef') would freeze the tests
that themselves use something similar ala use-ok('Zef')) and
also allows us to avoid accidently including `$*CWD/lib` when
checking for existence.

With the above in place we can use $dist.is-installed where
$dist is `Zef`, which will report a need for --force if its
already installed. Previously the `.is-installed` step was skipped
for `Zef` or `$dist.path ~~ $*CWD` which mimicked --force but did
not pass `:force` to `CU::R::Installation.install`