Make metacpan work again
Just the basics. Random segfaults from running the tar command
for some reason, but successfully installed `CSV::Parser`. Also
tried HTTP::UserAgent but the distros appear to be too old for
the tests to still pass. However it did fetch all the dependencies
and start testing them.

To test: edit resources/config.json so that :ContentStorage('cpan')
has "disabled" : 0, and set the other storages to "disabled" : 1