Try all extraction plugins if needed
This relates to the previous gnu tar / windows problem.
If an extraction plugin fails to extract, it will continue to
try with other available plugins in case one of them actually
does work.

Some sort of warning should eventually be added to make it clear
when this happens, but in general I think this is the expected
behavior.