Improved test handling
::Client.test() now takes a Candidate like .install, so ::CLI
just needs to call .candidates on the given items. As a bonus
`test XXX` now works on identities/urls and not just paths.

This also gets us one step closer to the next big addition
of a `Plan`/`Result` role for integrating work-to-do and
work-results into Candidates (currently just an anonymous role)