Plugin on/off: --$short-name; ex: --cpan
To enable or disable a plugin that is in the configuration file
you now pass --<short-name> or --<module> (or to disable:
--/<short-name> or --/<module>).

The intention is to allow easier use of metacpan over the previous
dynamic plugin enabling. So to search *only* cpan and not the defaults:
`zef --cpan --/p6c --/cached install CSV::Parser`

Ideally this gets improved to a user may set other fields of
the configuration and not just enable/disable.

`name` / `.plugin-id` have both been changed to `short-name` to
better match up with the CUR implementation