Fix --depsonly
This functionality was accidently removed in a recent commit. It
is now re-added, although it works slightly different to work
better with the recent split of:
`find-candidates` and `find-prereq-candidates`

So instead of adding the requested @identities to $exclude to be
ignored, it now just leaves out the @array containing the requested
candidates. This differs from before where requested and prereqs
were fetched in a single call, where ::Client itself needed to
filter out requested (which does not work anymore, as $excluded
will prevent `find-prereq-candidates` from even searching for
@requested (still need to fetch these so we can look at their
META6 depends fields)