Don't CI test with shrplib perl
Travis can't get their act together with OSX and perlbrew. It used to fail silently with the perlbrew commands and continue, but suddenly TravisCI fails to even parse the command (getting tripped up on the `.` of 5.20... wut?) and exits immediately with code 127. This means to properly CI test with such a perl that a bunch of duplicated config would need to be used for OSX. Due to this we now test with the default perl and have removed testing the install of Inline::Perl5