Fix git@ uris for reals
The original fix (not in a PR) was changed slightly and retested.
Retesting appeared to have worked, but really it was loading the
original fix's precomp file still. The modified fix in the first
PR missed a match part that became apparent after manually
deleting `~./perl6` and `zef/lib/.precomp`

Resolves #86 (2nd time's a charm)

zoffixznet++ for the detailed reports and persistence