Add temp shims for Distribution update
Zef::Distribution used to subclass Distribution, but the PR729
changes to rakudo have changed it into a unsubclassable role. This
adds some temporary shims to allow zef to continue functioning
on older rakudos as well as rakudos post PR729.

Yes this is a hacky mc hack hack. Soon Zef::Distribution will be
reworked entirely to make better use of PR729 changes, and a more
appropriate compatibility layer can be added at that time.