Workaround weird rakudo bug when setting config
Mixing in `method IO {$path}` to our %configuration would result in
`$path.IO` throw an error as if it was `Any.IO`. However `$path`
did infact contain a Str (and this still happend when coerced to
an IO::Path). So now we create an anonymous class which seems to
hold on to the correct values.

This bug only seems to be present when loading `Zef::CLI` via
`CompUnit::Repository::Tar`