Fix relative source path for reals
Before I was just using the CURI.install shim for rakudos with and
without the new Distribution interface. However the shim was
changed from the original PR to set a prefix to $*CWD; before
I was counting on passing a relative IO path object that would
stringify to a relative path while retaining its own CWD so copy()
Did The Right Thing.

Instead we just say fuck you old rakudo user and save absolute
paths *only* in this situation. Newer rakudos will save with the
relative path which will be expected going forward (this wasn't
expected one way or the other before, so its not really a fuck
you to old rakudo users)