Fix parsing of win abs paths
The parsing is only required for determining if a string is either
a path, uri, or identity. An identity is "Foo::Bar", but path and
uri may be considered the same in many cases. For our purposes the
only difference between path and uri is that path is a string
that is not checked by Zef::Utils::URI's file grammar before
marking it as a file (anything starting with / or .). On windows
such a path check would fail for absolute paths that include the
volume, so it gets picked up in a secondary check via
Zef::Utils::URI where it is prefixed with "file://" (if required)
and parsed with a grammar