Allow $HOME/.config/zef or XDG_CONFIG_HOME for config
This commit checks a list of path candidates, if XDG_CONFIG_HOME
is defined it checks that, otherwise $HOME/.config/zef.
If that file doesn't exist it checks the ones it checked previously.

I believe this is slightly faster now because it checks for the
file's existence before trying to parse it. It also will show
an specific error if the file doesn't exist versus if it doesn't
parse.

zef/issues/158