Fix generic cache store
* Copy files from store to cache

This wasn't needed before because we just pointed it at the cache
directory - eg. it didn't need to copy things into its own cache
because it just used where it was. However this meant adding
another LocalCache with config option :cache<.some-dir> would
only keep track of dists in an index and not actually cache
anything.

* update returns bool