Try to cleanup tmp stuff
If a fetch also does an extract which succeeds then its safe to
delete that path.

Unfortunately this will make git repos take longer to fetch when
updated, since the git repo itself will now be deleted after the
approriate files are in the store. But realistically we want
people to avoid using git repos as their source-url anyway, so
this is just another reason to use tarballs/cpan for releases.