Zef::Distribution: move BUILDALL -> TWEAK
Some of the details around BUILDALL have changed in recent versions of
rakudo for speed reasons.  Rather than adjusting our use of BUILDALL
here to handle the fact that it is now a multi, move over to using TWEAK
which seems to more-closely match what we are wanting to do here.

Without this change:

    % perl6 -Ilib bin/zef install .
    ===SORRY!=== Error while compiling /Users/aeruder/proj/rakudobuild/zef/lib/Zef/Distribution.pm6 (Zef::Distribution)
    Cannot have a multi candidate for 'BUILDALL' when an only method is also in the package 'Zef::Distribution'
    at /Users/aeruder/proj/rakudobuild/zef/lib/Zef/Distribution.pm6 (Zef::Distribution):48