Improve understanding of archive files
Previously it was assumed an archive contained the root prefix
directory as the first path from its listing.

However this isn't technically needed, and we can instead use
the parent directory of the shortest path that ends with
'/META6.json'.

This change allows a wider range of archive file configs to work by
implementing the above.