Fix zef update for dists that change auth
First, the 'update' command still needs to be redone more proper.
What this does is allow upgrading dists that adds an auth field
that is not present in the older dist that triggered the upgrade.
The proper way would be to use zef's .matches-spec api after
restructuring some of the looping constructs.

Resolves #197