Fix extra newlines in terminal output
We may receive more or less than a line of decoded data, so we
cannot naively assume that we can add a newline after each chunk
received.

However, this is likely slightly less performant when multiple
lines are received at once since we now `say` each line instead of
the entire chunk.