Tweak legacy builder logic
This works with (and without) the candidates PRs in rakudo,
and without temporarily rewriting a meta file (to write in
Build.pm to provides). Not sure why this has to be done in
such a specific way, but it suffices to say that "it just
needs to be this way for now", where "for now" is such a time
or soluton that works for installing both Inline::Perl5 and
Linenoise.