Add CURI as distribution source
This enables CURI to be a backend for zef.

This is still not complete. As it stands this commit will blow up
when doing e.g. `zef install Test` because the meta data from
installed modules does not have a distribution mapping.

Also see: https://github.com/ugexe/zef/issues/230
( although a smaller scoped fix may be used for this)