Explicitly ignore CORE modules
Normally we are protected by a .is-installed check for things like
Test, but some code paths ( such as `zef depends ...` ) do not use
the $skip-installed flag.

A better fix will be when Zef::Repository:: and
CompUnit::Repository:: are better integrated so instead of
ignoring Test / NativeCall / etc ( CORE ) we can grab its metadata
like everything else.

Resolves: #230