Make build process consistent with non-build dependencies
* Return to spawning in a new process -- this is so its easier to
set the libs available to the dependency. There is probably a nice
way to do this with a temp $*REPO to avoid spawning a new process.

* No longer install build dependencies differently than other
dependencies -- this allows --dry --serial etc to continue being
useful.

Essentially we want each set of dependencies to have an isolated
environment, but installing build dependencies prior to anything
else means its much more difficult to get reproducibility between
run.