Use unmodified projects.json mirror by default
The old url contains mangled meta data so older versions of zef can continue working with no impact. The new url allows more modern versions of zef to use more advanced dependency resolution. In the future the old urls can be made to contain unmodified data, and the new url deleted ( which will work because it will fallback to the other mirrors )