Add timeout functionality to various phases (#255)
Allows a --fetch-timeout option which will end any downloading
once X number of seconds have elapsed since the download was
initially attempted. Defaults to 10 minutes for the general case, and
3 minutes when fetching the ecosystem. There is currently no way to
override the ecosystem timeout yet though.

Also allows --build-timeout, --extract-timeout, and --test-timeout
which follow similar rules to the above but default to 60 minutes.