Add a better error message for invalid 'resources'
When 'resources' field has something other than `Str` to safely match
against, e.g. a Pair, it threw with `Invocant of method 'match' must
be a type object of type 'Any', not an object instance of type
'Pair'.`, which is not helpful.  Here we explicitly check for each
resource item to be a Str and provide an error message that is easier
to understand.