Refactor logging system to accomodate parallelization
With parallelization the output would previously be confusing
because you would not know which lines belonged to a specific
distribution being processed. The main changes this introduces
involve passing the $candi object around to the strategy classes,
so they in turn can emit them on the logger (allowing the logger
to print the name of the distribution a line resulted from).