Rework META.info support
We removed the use of `-Ilib` but that broke support for the
deprecated META.info distributions. Since there are quite a
few of these and they have unaddressed PRs to change the filename
we will have to continue supporting them some other way. That
"other way" is to just create a META6.json from the META.info
(so CURFS can understand it without the `-Ilib`).