Fix parsing of auths from identities containing < or >
Previously searching for a name of a module (not a distribution
name) any auth that would contain a < or > would cause any
matching to fail. This fixes the grammar to properly parse
such characters in an identity string, as well as changing
how identities are constructed for provides to avoid parsing
from a string (instead setting the attributes directly).

Resolves #327