Use exact versions for plugins
While the regular `use Foo` imports were changed to use exact
versions, there was still the issue of plugins (which used
the version-less `require Foo` syntax). This changes the plugin
code to take versions into account (defaulting to the executing
zef version as a default for anything provided by the zef
distribution). Additionally the plugin role was moved into its
own file so as not to bring a dependency into lib/Zef.pm6