Notify when backends are not configured (#339)
When failing to find a backend to fetch a module we give an error message stating as much. However, this logic was missing when fetching something other than a module (like the ecosystem package list). This adds the backend listing error message when fetching the e.g. ecosystem package list fails.