Index distributions by short name in ::Ecosystems
Searching for given distributions previous involved iterating over
every single distribution in the ecosystem to see if its name and
version parts matched. This adds indexing of distributions based on
their short name, so when making the actual query we can iterate
over only those distributions indexed on the short name of the
requested identity.

This brings an install of HTTP::UserAgent from 160s to less than
20s on the system I'm testing on.