Refactor/comment Fetcher|Extractor|Builder|Tester|Installer adapters
This refactors the various adapters slightly to use typing and to
make the code between similar adapters as similar as possible.
Functionally little to nothing should change as a result fo these
changes.

This also adds comments to the various adapters to assist users in
learning how to make or configure their own extensions. Perhaps
in the future this can all be turned into pod6 (which would be
a good low hanging fruit).