Fix spec comparison to work with uneven version strings
Long ago there was some additional code that mangled the version
string in such a way as to make the original comparison work. That
code was probably removed when part of the original issue it solved
was fixed in rakudo; however there are situations where this code
is still required (basically depending on being able to not care
which side each version is for a smartmatch). This returns the
original logic for balancing the version length before comparison.

Resolves #379