Fix erroneous deduplication of dependencies (#391)
The previous sorting logic abused grouping on a distributions name
which doesn't work so well with things like requiring two versions
of the same module. This change simplifies the sorting logic and
fixes the aforementioned issue by grouping on the `.as` value,
allowing e.g. Foo:ver<1> and Foo:ver<1.1+> to correctly be grouped
separately. And since those example dependencies may ultimately
resolve to the same distribution we also add a uniqueness filter
on the long name identity.

Resolves #382