Fix issue when no candidates are found
Previously this could give the error No such method 'as' for
invocant of type 'Any' such as when it failed to find a native
dependency that wasn't already installed. This change returns
the original behavior where zef presents a message telling the
user it cannot find the dependency and any suggestions it has.