Make Zef::Repository.update return Nil
The documentation for Z:R.update stated the return value was likely
to be removed in the future, and now it is so. This makes its behavior
consistent with the interface for all the Z:R:*.update adapters.

This also makes the table output of `zef update` only occur when
using --verbose or --debug. This makes `zef update` take ~2s and
`zef update --verbose` take ~20s instead of both taking ~20s. The
actual reason its so slow is because there isn't a way to count
the number of distributions without creating the distribution
objects first (which is hard to avoid since creating the object
automatically greps out meta data that isnt valid in certain ways).
Since having a low/no output for `zef update` makes sense anyway
this can stand in as a workaround in the mean time.