Maintain config order preference of ecosystems
When searching for a specific distribution it shouldn't matter
which backend it comes from, and hence `race` was originally used.
However this doesn't provide any benefit and comes at the cost of
reproduciblity. This changes the search code to use `hyper` instead
of `race` to maintain the backend order from the config such that
two runs will always draw from the same backend / ecosystem.