Don't bypass interface by extracting into cache directory (#411)
When adding distributions to the local cache we previously would
e.g. extract the tar file to $cache-dir. However this can be
problematic because the extractors are not aware of the required
file structure of the $cache-dir. This can lead to consumers of
$cache-dir to do funny thing.

This also adds a check to the cache modules store method so that
it won't try to store something that is already stored in the cache.

Both of the above fix an issue where distributions in a tar archive
without a directory prefix in the archive file would lead to the
cache storing additionally copies of itself within itself.