Allow dependency confusion mitigation by grouping ecosystems (#400)
Currently all ecosystems are given equal considerations when making
a recommendation for a namespace. This change changes the array of
recommendation managers to an array of an array of recommendation
managers. This allows a setup like `[[Darkpan],[fez,p6c]]` where
if `Foo` is found in Darkpan it will always be chosen, even if fez
or p6c have a `Foo` with a higher version.