Use the cache less often when querying for distributions
Before v0.13.0 when the repositories in the config were not grouped
the cache could easily act as a universal alternative for given
distribution names. Now that ecosystems are grouped it could only
act as an alternative for the ecosystems it is explicitly grouped
with, but doing so also shadows lower tier repos in a way many
users may not expect.

This moves the cache to the very last possible ecosystem to query
from, relegating it to mostly serve as a store of things you may
have downloaded (in case the release/branch disappears or changes)
rather than a means to save bandwidth.