Support increase verbosity in TAP plugin
The previous TAP plugin required capturing $*OUT/$*ERR from TAPs
output which complicated showing that data to the user (such as
prefixing each line with i.e. '[zef] ...'). This uses new changes
to TAP to integrate with stdout and stderr Suppliers, allowing
output to be handled similarly as most other plugins.

This also sets HARNESS_VERBOSITY if --debug is passed which will
e.g. show individual test results and not just the result of
individual files.