Use short name for name option
The short name better lets the user know how to enable/disable something from the output they see, but the CLI output still uses this name value when it should be using the short name. In the future this name field could be better integrated and the other plugins updated to have a more descriptive name.