Resolve path created from relative path parts
When creating a path from relative path parts its possible to end
up with a path like /foo/bar/.. which is problematic because
"/foo/bar/..".IO.parent returns "/foo/bar" not "/foo". This adds
a call to .resolve on one such IO::Path to give the later instead
of the former, fixing a potential erroneous mkdir call on the
path of a file instead of the parent directory of said file.