Pass loggers through functions instead of initializing with object (#497)
At one time the various plugins would be created anew for each use,
but later on we started caching those instances. That change led to
Supply reuse that showed itself as duplicate messages on any use
of said plugin beyond the first.

This passes in the loggers through the function calls such that
we can create new ones before we call the function. This avoids
the Supply reuse issue mentioned earlier.

Fixes #496