Make probe methods thread safe (#504)
Previously various probe methods used `state`, which is not thread
safe. Since we preload all the plugins it didn't need to be thread
safe. This updates various probe methods to be thread safe despite
the above comment only because it is technically more correct and
potentially makes updating the plugin system a bit easier.