Avoid printing to stdout/stderr inside repository
All other parts of zef send output to stdout/stderr through a
logging mechanism. The Repository (and Repository plugins) were
the only plugins that were not already doing this. This updates
the aforementioned Repository related code to using the idiomatic
logging mechanism to send terminal output (specifically the output
to tell the user when an ecosystem has been updated).