Fix tar detection on OpenBSD
The tar command on OpenBSD does not have a --help or -h flag,
which we relied on for checking the exit code as well as determining
if the tar command on the system supported .zip files. This adds
a special case for the tar plugin on OpenBSD so that it can be
detected properly.