Use exact versions for imports
Because of the way CUR resolves versions it is possible for a v2
Foo::Bar to load e.g. v1 of Foo::Bar::Baz if the later exists in
a repo earlier than the v2 (i.e. 'home' is before 'site'). This
is particularly noticible with `zef`, where old versions from
rakudo star get used by the *new* bin/zef script in PATH.

This adds a hard version requirement on the various Zef module
imports that is based on whatever version is current in the
META6.json file.

Note this is still missing a critical piece -- a way for bin/zef
to say `use Zef::CLI:ver($?DISTRIBUTION.meta<version>)` which
itself doesn't work because $?DISTRIBUTION / %?RESOURCES are not
(yet...) available to bin scripts.