Allow cache directory to be set from config
Zef::Repository classes can be configured in resources/config.json
to pass in options for class initialization. Previously this
could not be done in json because `cache` wanted an `IO::Path`,
but a user can't supply that in JSON. This updates the type of
the `cache` path attribute to `IO()` so a string path gets coerced
into an `IO::Path`.