Add 'ver' fallback to 'version'
This adds to the previous fallback commit. When -Ilib is used, it
creates a `Distribution::Hash` using the 'ver' key. That means
$?DISTRIBUTION.meta<version> would not return a value when -Ilib
is used. Admittedly the initial import statement in each file is
now obnoxiously large. In the future we should probably do something
in rakudo to make 'version' show up, since telling users to start
using 'ver' instead would seem much harder.