Remove prove test plugin
One reason this is being removed is in the spirit of dogfooding:
if one wants to use a test harness they should use the Raku one
which allows us to get back structured information.

Another reason is this plugin makes it difficult to pass in the
includes paths. Previous we were just setting RAKULIB, but that
is causing issues since we changed to using a staging installation
workflow.