Pass includes paths on command line instead of setting RAKULIB
Previously we set the includes paths of all the dependencies that
had been downloaded (but not yet installed) to RAKULIB to be
defensive about hitting a command length limitation if we were
to instead pass those paths in via -I. Now that we use a staging
installation workflow we only ever have one path to include, so
we don't have to worry about hitting any command length limitation.
The way we changed RAKULIB after the staging had occured also broke
installations for users that had RAKULIB set to something already,
so we now have a reason to stop setting RAKULIB.

This passes in test includes via `-I/some/path` instead of via
`RAKULIB="/some/path,$RAKULIB"`, and fixes installations for users
who have `RAKULIB` set when running zef.