Avoid warnings when ecosystem contains invalid distribution name
Previously it was possible (although extremely unlikely) that a
distribution could exist in an ecosystem without a name field,
which could lead to warnings when we attempt to stringify it. A
similar thing could also occur when an invalid long name is used
in the provides field, such that we fail to parse the name. This
ignores such distributions during the indexing process.

Fixes #543