Allow positional arguments passed in through STDIN via '-'
Previously there was no way to pass in arguments to zef via stdin,
which can be useful for large argument lists or when piping from
a list of modules. This updates zef to treat '-' for zef install
and update to read the data from stdin and treat them as positional
arguments.