Make sure there is a path part to create path from
Previously a META6.json containing a resource entry of `libraries/`
would cause an error when we tried to create an IO path from the
first child part (which doesn't exist in the given entry). This
fixes the regex for matching "libraries" by making sure an expected
path part actually exists.