Optimize use of Supply
Previously we were creating a new supply for each ecosystem that
was being updated, as well as only closing the stderr (and not
stdout) supply. This updates this logic to match what we do in
e.g. Zef::Fetch by reusing the same two supplies for every
ecosystem as well as explicitly closing them both when we are done.